Refactor dataservice derive functionality.
Creating a new DataDeriveUtil class in dataservice package and moving
code related to deriving datasets from service manager class to this new
class.
